requirements.txt
streamlit
streamlit-webrtc
fer
whisper
torch
transformers
Pillow
av
numpy